full CAN API refactor
